Warning:  length of some xref entries is not equal to 20 bytes.
If  the  file’s  end-of-line  marker  is  a  single  character (either a carriage return or a line feed), it is preceded by a single space; if  the  marker  is  2  characters  (both  a  carriage  return  and  a  line  feed),  it  is  not  preceded  by  a  space.  Thus,  the  overall  length  of  the  entry  is  always  exactly  20  bytes.